Breakdown CreateHook component into smaller components for re-rendering optimization
